Move getDailyMenuURL to another folder
